refactor: 将 app/core/module.py 拆分为 app/core/module 包
按职责拆分单文件 module.py 为包，对外公共 API 零改动（几十处 from app.core.module
import ModuleManager 不变）：

- loader.py   : 原 module_loader.py 的扫描/发现（ModuleHelper / FilterFuncType）
- manager.py  : ModuleManager 注册/生命周期
- contract.py : verify_module_contract / _verify_typed_contract / verify_*_contract 家族
- __init__.py : PEP 562 懒加载再导出 ModuleManager / ModuleType / verify_*_contract

兼容性：
- ModuleManager.verify_*_contract 以 staticmethod 别名保留，plugin_manager 与测试零改动；
- app/core/module_loader.py 转为指向 app.core.module.loader 的兼容垫片，
  app.helper.module 垫片链与 app.core.module_loader 历史路径解析到同一批对象；
- __init__ 懒加载保住「import app.core.module.loader 不触发 manager 重依赖」的 S3 解耦不变量
  （ModuleHelper 被 filemanager/indexer/workflow 在 import 期使用）。

全量测试 1713 passed / 19 skipped。
